+++ packages/site/src/content/AtlantisThemeContext/index.tsx
@@ -1,4 +1,6 @@
-import AtlantisThemeContextContent from "./AtlantisThemeContext.stories.mdx";
+import AtlantisThemeContextContent, {
+  toc,
+} from "./AtlantisThemeContext.stories.mdx";
 import Props from "./AtlantisThemeContext.props.json";
 import MobileProps from "./AtlantisThemeContext.props-mobile.json";
 import { ContentExport } from "../../types/content";
@@ -6,6 +8,7 @@ import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <AtlantisThemeContextContent />,
+  toc,
   props: Props,
   mobileProps: MobileProps,
   component: {