+++ packages/site/src/content/EmptyState/index.tsx
@@ -1,10 +1,11 @@
-import Content from "./EmptyState.stories.mdx";
+import Content, { toc } from "./EmptyState.stories.mdx";
 import MobileProps from "./EmptyState.props-mobile.json";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
+  toc,
   mobileProps: MobileProps,
   component: {
     mobileElement: `<EmptyState