+++ packages/site/src/content/FeatureSwitch/index.tsx
@@ -1,10 +1,11 @@
-import Content from "./FeatureSwitch.stories.mdx";
+import Content, { toc } from "./FeatureSwitch.stories.mdx";
 import Props from "./FeatureSwitch.props.json";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
+  toc,
   props: Props,
   component: {
     element: `const [featureEnabled, setFeatureEnabled] = useState(true);