+++ packages/site/src/content/InlineLabel/index.tsx
@@ -1,10 +1,11 @@
-import InlineLabelContent from "./InlineLabel.stories.mdx";
+import InlineLabelContent, { toc } from "./InlineLabel.stories.mdx";
 import Props from "./InlineLabel.props.json";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <InlineLabelContent />,
+  toc,
   props: Props,
   component: {
     element: `<InlineLabel>Default</InlineLabel>`,