+++ packages/site/src/content/RecurringSelect/index.tsx
@@ -1,10 +1,11 @@
-import Content from "./RecurringSelect.stories.mdx";
+import Content, { toc } from "./RecurringSelect.stories.mdx";
 import Props from "./RecurringSelect.props.json";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
+  toc,
   props: Props,
   component: {
     element: `const [rule, setRule] = useState<RecurrenceRule>({