+++ packages/site/src/content/Text/index.tsx
@@ -1,4 +1,4 @@
-import Content from "./Text.stories.mdx";
+import Content, { toc } from "./Text.stories.mdx";
 import Props from "./Text.props.json";
 import MobileProps from "./Text.props-mobile.json";
 import Notes from "./TextNotes.mdx";
@@ -7,6 +7,7 @@ import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
+  toc,
   props: Props,
   mobileProps: MobileProps,
   component: {