+++ packages/site/src/content/ThumbnailList/index.tsx
@@ -1,10 +1,11 @@
-import Content from "./ThumbnailList.stories.mdx";
+import Content, { toc } from "./ThumbnailList.stories.mdx";
 import MobileProps from "./ThumbnailList.props-mobile.json";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
+  toc,
   props: MobileProps,
   component: {
     mobileElement: `<ThumbnailList