+++ packages/components-native/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.98.1](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.98.0...@jobber/components-native@0.98.1) (2026-03-13)
+
+**Note:** Version bump only for package @jobber/components-native
+
+
+
+
+
+# [0.98.0](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.97.1...@jobber/components-native@0.98.0) (2026-03-12)
+
+
+### Features
+
+* **components-native:** Replace react-native-keyboard-aware-scroll-view with react-native-keyboard-controller ([#2901](https://github.com/GetJobber/atlantis/issues/2901)) ([f2adedb](https://github.com/GetJobber/atlantis/commit/f2adedb36f5d7e5fa5d8249c0bd6fa251c9aa32e))
+
+
+
+
+
 ## [0.97.1](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.97.0...@jobber/components-native@0.97.1) (2026-03-03)